requirements.txt
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
schedule>=1.2.0
click>=8.1.7
aiohttp>=3.9.0
lxml>=5.1.0
beautifulsoup4>=4.12.3
sqlalchemy>=2.0.0
pydantic>=2.6.0
tenacity>=8.2.3
